pyproject.toml
[project]
name = "ai-research-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "dvc>=3.59.1",
    "faiss-cpu>=1.11.0",
    "langchain-community>=0.3.22",
    "mlflow>=2.22.0",
    "mypy>=1.15.0",
    "pymupdf>=1.25.5",
    "pyspark>=3.5.5",
    "ruff>=0.11.7",
    "sentence-transformers>=4.1.0",
]
